requirements.txt
requests==2.28.1
termcolor==2.2.0
python-dotenv==0.19.2